Circle warping fixed
Circles appear to warp in uneven x and y scales after being assigned as a child to the box object, which also has different local x and y scales.

Instead a new empty GameObject is created to represent the box (1,1,1 local scale).

Original box object is now just a DisplayBox - for displaying where the box is only